requirements.txt
# Web Framework and Server
fastapi>=0.109.2
uvicorn>=0.27.1
python-multipart>=0.0.19

# Database and ORM
sqlalchemy>=2.0.27
psycopg2-binary>=2.9.9
alembic>=1.13.1

# Data Validation
pydantic>=2.6.1,<3.0.0
pytest-asyncio>=0.23.0

# Environment and Configuration
python-dotenv>=1.0.1

# Development and Testing
pytest>=8.0.2
httpx>=0.26.0
black==25.1.0
